[SERVER_DEV66] Modify "Map to Set" //DEV_66
 기존에 UserDataManager에서 사용하던 Custom_Map_Tree 코드를 수정하여 Custom_Set_Tree 생성 및 적용. (Key String Memory overhead)

UserDataManager에서, UserData는 Custom_Set_Tree 사용
NickNameCont는 Custom_Map_Tree ( nickName - wstring, id - string) 사용

추후 id를 string에서 int혹은 long 변경에 대한 유연성 갖춤
